# backend/cmd/api/routes.go
package api

import (
	"transx/cmd/api/handlers"
	authdto "transx/internal/modules/auth/application/dto"
	notifdto "transx/internal/modules/notification/application/dto"
	transferdto "transx/internal/modules/transfer/application/dto"
	walletdto "transx/internal/modules/wallet/application/dto"

	"github.com/gofiber/fiber/v2"
	"github.com/oaswrap/spec/adapter/fiberopenapi"
	"github.com/oaswrap/spec/option"
)

// RegisterRoutes wires the auth handlers onto the Fiber app for the running auth
// service. The spec router groups under /api/v1 to match the gateway prefix.
func RegisterRoutes(app *fiber.App, authH *handlers.AuthHandler) {
	RegisterAuthRoutes(fiberopenapi.NewRouter(app), authH)
}

// RegisterWalletRoutes wires the wallet handlers onto the Fiber app for the
// running wallet service. Kept separate from RegisterRoutes so each binary only
// registers the routes it actually serves (a shared registrar would force a nil
// handler for the other service and panic at request time).
func RegisterWalletRoutes(app *fiber.App, walletH *handlers.WalletHandler) {
	registerWalletRoutes(fiberopenapi.NewRouter(app), walletH)
}

// RegisterTransferRoutes wires the transfer handlers onto the Fiber app for the
// running transfer service. Kept separate from RegisterWalletRoutes so each
// binary only registers the routes it actually serves (a shared registrar
// would force a nil handler for the other service and panic at request time).
func RegisterTransferRoutes(app *fiber.App, transferH *handlers.TransferHandler) {
	registerTransferRoutes(fiberopenapi.NewRouter(app), transferH)
}

// RegisterAllRoutesForSpec registers every route with nil handlers so the
// OpenAPI exporter can emit the full spec without wiring real dependencies. Nil
// handlers are safe here because the exporter never invokes them.
func RegisterAllRoutesForSpec(r fiberopenapi.Router) {
	RegisterAuthRoutes(r, nil)
	registerWalletRoutes(r, nil)
	registerTransferRoutes(r, nil)
	registerInboxRoutes(r, nil)
}

// RegisterInboxRoutes wires the inbox-service routes onto the Fiber app.
func RegisterInboxRoutes(app *fiber.App, inboxH *handlers.InboxHandler) {
	registerInboxRoutes(fiberopenapi.NewRouter(app), inboxH)
}

// registerInboxRoutes wires the inbox routes onto the spec router.
// A nil handler registers the route for spec export only.
func registerInboxRoutes(r fiberopenapi.Router, inboxH *handlers.InboxHandler) {
	v1 := r.Group("/api/v1/inbox")

	var getUnreadCount, listInbox, getInbox, readAll fiber.Handler
	if inboxH != nil {
		getUnreadCount = inboxH.GetUnreadCount
		listInbox = inboxH.ListInbox
		getInbox = inboxH.GetInbox
		readAll = inboxH.ReadAll
	}

	// Static paths before /:id so /unread-count and /read-all are not caught by the
	// parameterized route.
	v1.Get("/unread-count", getUnreadCount).With(
		option.Tags("inbox"),
		option.OperationID("getInboxUnreadCount"),
		option.Summary("Return the caller's unread inbox item count"),
		option.Response(fiber.StatusOK, new(notifdto.UnreadCountResponse)),
		option.Response(fiber.StatusUnauthorized, new(handlers.ErrorResponse)),
	)

	v1.Post("/read-all", readAll).With(
		option.Tags("inbox"),
		option.OperationID("markAllInboxRead"),
		option.Summary("Mark all of the caller's unread inbox items as read"),
		option.Response(fiber.StatusOK, new(notifdto.ReadAllResponse)),
		option.Response(fiber.StatusUnauthorized, new(handlers.ErrorResponse)),
	)

	v1.Get("/", listInbox).With(
		option.Tags("inbox"),
		option.OperationID("listInbox"),
		option.Summary("List the caller's inbox items, newest first, paginated"),
		option.Request(new(notifdto.ListInboxQuery)),
		option.Response(fiber.StatusOK, new(notifdto.InboxListResponse)),
		option.Response(fiber.StatusUnauthorized, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusInternalServerError, new(handlers.ErrorResponse)),
	)

	v1.Get("/:id", getInbox).With(
		option.Tags("inbox"),
		option.OperationID("getInboxItem"),
		option.Summary("Get one inbox item. If unread it is automatically marked as read."),
		option.Response(fiber.StatusOK, new(notifdto.InboxItemResponse)),
		option.Response(fiber.StatusBadRequest, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusUnauthorized, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusNotFound, new(handlers.ErrorResponse)),
	)
}

// RegisterAuthRoutes wires the auth-service routes: login/refresh/logout + the
// ForwardAuth check endpoint. Passing a nil handler registers the route for
// spec export only (no live handler attached).
func RegisterAuthRoutes(r fiberopenapi.Router, authH *handlers.AuthHandler) {
	v1 := r.Group("/api/v1")

	var login, refresh, logout, session, check fiber.Handler
	if authH != nil {
		login = authH.Login
		refresh = authH.Refresh
		logout = authH.Logout
		session = authH.Session
		check = authH.Check
	}

	v1.Post("/login", login).With(
		option.Tags("auth"),
		option.OperationID("login"),
		option.Summary("Authenticate with email and password; returns access + refresh tokens (JSON)"),
		option.Request(new(authdto.LoginCommand), option.ContentRequired()),
		option.Response(fiber.StatusOK, new(authdto.LoginResponse)),
		option.Response(fiber.StatusBadRequest, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusUnauthorized, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusForbidden, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusInternalServerError, new(handlers.ErrorResponse)),
	)

	v1.Post("/refresh", refresh).With(
		option.Tags("auth"),
		option.OperationID("refresh"),
		option.Summary("Rotate refresh token and return a new access + refresh pair"),
		option.Request(new(authdto.RefreshCommand), option.ContentRequired()),
		option.Response(fiber.StatusOK, new(authdto.LoginResponse)),
		option.Response(fiber.StatusUnauthorized, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusForbidden, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusInternalServerError, new(handlers.ErrorResponse)),
	)


	v1.Post("/session", session).With(
		option.Tags("auth"),
		option.OperationID("session"),
		option.Summary("Validate refresh token without rotating it (RR BFF auth-gate)"),
		option.Request(new(authdto.RefreshCommand), option.ContentRequired()),
		option.Response(fiber.StatusNoContent, nil),
		option.Response(fiber.StatusUnauthorized, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusBadRequest, new(handlers.ErrorResponse)),
	)

	v1.Post("/logout", logout).With(
		option.Tags("auth"),
		option.OperationID("logout"),
		option.Summary("Revoke refresh session (JSON refreshToken)"),
		option.Request(new(authdto.RefreshCommand)),
		option.Response(fiber.StatusNoContent, nil),
		option.Response(fiber.StatusForbidden, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusInternalServerError, new(handlers.ErrorResponse)),
	)

	// ForwardAuth backend used by Traefik. On success it returns 200 with the
	// X-User-Id response header; on an invalid/missing token it returns 401.
	v1.Get("/check", check).With(
		option.Tags("auth"),
		option.OperationID("check"),
		option.Summary("ForwardAuth check: verify bearer token, echo X-User-Id"),
		option.Response(fiber.StatusOK, nil),
		option.Response(fiber.StatusUnauthorized, new(handlers.ErrorResponse)),
	)
}

// registerWalletRoutes wires the wallet-service routes onto the spec router.
// A nil handler registers the route for spec export only.
func registerWalletRoutes(r fiberopenapi.Router, walletH *handlers.WalletHandler) {
	v1 := r.Group("/api/v1")

	var createAccount, getAccount, lookupAccount, listAccounts fiber.Handler
	if walletH != nil {
		createAccount = walletH.CreateAccount
		getAccount = walletH.GetAccount
		lookupAccount = walletH.LookupAccount
		listAccounts = walletH.ListAccounts
	}

	v1.Post("/accounts", createAccount).With(
		option.Tags("wallet"),
		option.OperationID("createAccount"),
		option.Summary("Create a wallet account for the caller"),
		option.Request(new(walletdto.CreateAccountCommand), option.ContentRequired()),
		option.Response(fiber.StatusCreated, new(walletdto.AccountResponse)),
		option.Response(fiber.StatusBadRequest, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusUnauthorized, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusInternalServerError, new(handlers.ErrorResponse)),
	)

	v1.Get("/accounts", listAccounts).With(
		option.Tags("wallet"),
		option.OperationID("listAccounts"),
		option.Summary("List the caller's wallet accounts, paginated, with optional currency and status filters"),
		option.Request(new(walletdto.ListAccountsQuery)),
		option.Response(fiber.StatusOK, new(walletdto.AccountListResponse)),
		option.Response(fiber.StatusBadRequest, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusUnauthorized, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusInternalServerError, new(handlers.ErrorResponse)),
	)

	v1.Get("/accounts/:accountType/:accountRef", lookupAccount).With(
		option.Tags("wallet"),
		option.OperationID("lookupAccount"),
		option.Summary("Look up an internal or external account for transfer beneficiary validation"),
		option.Response(fiber.StatusOK, new(walletdto.AccountLookupResponse)),
		option.Response(fiber.StatusBadRequest, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusUnauthorized, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusNotFound, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusBadGateway, new(handlers.ErrorResponse)),
	)

	v1.Get("/accounts/:accountRef", getAccount).With(
		option.Tags("wallet"),
		option.OperationID("getAccount"),
		option.Summary("Get a wallet account balance by its accountRef (ACC- + ULID)"),
		option.Response(fiber.StatusOK, new(walletdto.AccountResponse)),
		option.Response(fiber.StatusUnauthorized, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusNotFound, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusInternalServerError, new(handlers.ErrorResponse)),
	)
}

// registerTransferRoutes wires the transfer-service routes onto the spec
// router. A nil handler registers the route for spec export only.
func registerTransferRoutes(r fiberopenapi.Router, transferH *handlers.TransferHandler) {
	v1 := r.Group("/api/v1")

	var createTransfer, getTransfer, listTransfers fiber.Handler
	if transferH != nil {
		createTransfer = transferH.CreateTransfer
		getTransfer = transferH.GetTransfer
		listTransfers = transferH.ListTransfers
	}

	v1.Post("/transfers", createTransfer).With(
		option.Tags("wallet"),
		option.OperationID("createTransfer"),
		option.Summary("Create a transfer (idempotent via Idempotency-Key). INTERNAL needs toAccountRef (an ACC- account ref); EXTERNAL omits it or carries a free-text beneficiary id (provider is set from server config)."),
		option.Request(new(transferdto.CreateTransferCommand), option.ContentRequired()),
		option.Response(fiber.StatusAccepted, new(transferdto.TransferResponse)),
		option.Response(fiber.StatusBadRequest, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusUnauthorized, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusForbidden, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusConflict, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusUnprocessableEntity, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusInternalServerError, new(handlers.ErrorResponse)),
	)

	v1.Get("/transfers", listTransfers).With(
		option.Tags("wallet"),
		option.OperationID("listTransfers"),
		option.Summary("List the caller's transfers, paginated, with optional status and accountRef filters"),
		option.Request(new(transferdto.ListTransfersQuery)),
		option.Response(fiber.StatusOK, new(transferdto.TransferListResponse)),
		option.Response(fiber.StatusBadRequest, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusUnauthorized, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusInternalServerError, new(handlers.ErrorResponse)),
	)

	v1.Get("/transfers/:transferId", getTransfer).With(
		option.Tags("wallet"),
		option.OperationID("getTransfer"),
		option.Summary("Get a transfer by its business reference (ETN- for EXTERNAL, ITN- for INTERNAL, followed by a ULID)"),
		option.Response(fiber.StatusOK, new(transferdto.TransferResponse)),
		option.Response(fiber.StatusBadRequest, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusUnauthorized, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusNotFound, new(handlers.ErrorResponse)),
		option.Response(fiber.StatusInternalServerError, new(handlers.ErrorResponse)),
	)
}