# backend/openapi.yaml
openapi: 3.1.0
info:
  title: transx API
  description: transx API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /accounts:
    get:
      tags:
        - wallet
      summary: List the caller's wallet accounts, paginated, with optional currency and status filters
      description: List the caller's wallet accounts, paginated, with optional currency and status filters
      operationId: listAccounts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
        - name: currency
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DtoAccountListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
    post:
      tags:
        - wallet
      summary: Create a wallet account for the caller
      description: Create a wallet account for the caller
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DtoCreateAccountCommand'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DtoAccountResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
  /accounts/{accountRef}:
    get:
      tags:
        - wallet
      summary: 'Get a wallet account balance by its accountRef (ACC- + ULID)'
      description: 'Get a wallet account balance by its accountRef (ACC- + ULID)'
      operationId: getAccount
      parameters:
        - name: accountRef
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DtoAccountResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
  /accounts/{accountType}/{accountRef}:
    get:
      tags:
        - wallet
      summary: Look up an internal or external account for transfer beneficiary validation
      description: Look up an internal or external account for transfer beneficiary validation
      operationId: lookupAccount
      parameters:
        - name: accountType
          in: path
          required: true
          schema:
            type: string
        - name: accountRef
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DtoAccountLookupResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '502':
          description: Bad Gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
  /check:
    get:
      tags:
        - auth
      summary: 'ForwardAuth check: verify bearer token, echo X-User-Id'
      description: 'ForwardAuth check: verify bearer token, echo X-User-Id'
      operationId: check
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
  /inbox:
    get:
      tags:
        - inbox
      summary: List the caller's inbox items, newest first, paginated
      description: List the caller's inbox items, newest first, paginated
      operationId: listInbox
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DtoInboxListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
  /inbox/read-all:
    post:
      tags:
        - inbox
      summary: Mark all of the caller's unread inbox items as read
      description: Mark all of the caller's unread inbox items as read
      operationId: markAllInboxRead
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DtoReadAllResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
  /inbox/unread-count:
    get:
      tags:
        - inbox
      summary: Return the caller's unread inbox item count
      description: Return the caller's unread inbox item count
      operationId: getInboxUnreadCount
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DtoUnreadCountResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
  /inbox/{id}:
    get:
      tags:
        - inbox
      summary: Get one inbox item. If unread it is automatically marked as read.
      description: Get one inbox item. If unread it is automatically marked as read.
      operationId: getInboxItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DtoInboxItemResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
  /login:
    post:
      tags:
        - auth
      summary: Authenticate with email and password; returns access + refresh tokens (JSON)
      description: Authenticate with email and password; returns access + refresh tokens (JSON)
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DtoLoginCommand'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DtoLoginResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
  /logout:
    post:
      tags:
        - auth
      summary: Revoke refresh session (JSON refreshToken)
      description: Revoke refresh session (JSON refreshToken)
      operationId: logout
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DtoRefreshCommand'
      responses:
        '204':
          description: No Content
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
  /refresh:
    post:
      tags:
        - auth
      summary: Rotate refresh token and return a new access + refresh pair
      description: Rotate refresh token and return a new access + refresh pair
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DtoRefreshCommand'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DtoLoginResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
  /session:
    post:
      tags:
        - auth
      summary: Validate refresh token without rotating it (RR BFF auth-gate)
      description: Validate refresh token without rotating it (RR BFF auth-gate)
      operationId: session
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DtoRefreshCommand'
        required: true
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
  /transfers:
    get:
      tags:
        - wallet
      summary: List the caller's transfers, paginated, with optional status and accountRef filters
      description: List the caller's transfers, paginated, with optional status and accountRef filters
      operationId: listTransfers
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
        - name: status
          in: query
          schema:
            type: string
        - name: accountRef
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DtoTransferListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
    post:
      tags:
        - wallet
      summary: 'Create a transfer (idempotent via Idempotency-Key). INTERNAL needs toAccountRef (an ACC- account ref); EXTERNAL omits it or carries a free-text beneficiary id (provider is set from server config).'
      description: 'Create a transfer (idempotent via Idempotency-Key). INTERNAL needs toAccountRef (an ACC- account ref); EXTERNAL omits it or carries a free-text beneficiary id (provider is set from server config).'
      operationId: createTransfer
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - fromAccountRef
                - amount
                - currency
                - message
              properties:
                amount:
                  type: string
                currency:
                  type: string
                fromAccountRef:
                  type: string
                message:
                  type: string
                toAccountRef:
                  type: string
                transferType:
                  type: string
        required: true
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DtoTransferResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
  /transfers/{transferId}:
    get:
      tags:
        - wallet
      summary: 'Get a transfer by its business reference (ETN- for EXTERNAL, ITN- for INTERNAL, followed by a ULID)'
      description: 'Get a transfer by its business reference (ETN- for EXTERNAL, ITN- for INTERNAL, followed by a ULID)'
      operationId: getTransfer
      parameters:
        - name: transferId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DtoTransferResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
components:
  schemas:
    DtoAccountListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DtoAccountResponse'
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
    DtoAccountLookupResponse:
      type: object
      properties:
        accountRef:
          type: string
        currency:
          type: string
        holderName:
          type: string
        status:
          type: string
    DtoAccountResponse:
      type: object
      properties:
        accountRef:
          type: string
        availableBalance:
          type: string
        currency:
          type: string
        holdBalance:
          type: string
        status:
          type: string
    DtoCreateAccountCommand:
      type: object
      required:
        - currency
      properties:
        currency:
          type: string
        name:
          type: string
    DtoInboxItemResponse:
      type: object
      properties:
        body:
          type: string
        createdAt:
          type: string
        id:
          type: string
        readAt:
          type:
            - string
            - 'null'
        title:
          type: string
        transferId:
          type: string
        type:
          type: string
    DtoInboxListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DtoInboxItemResponse'
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
    DtoLoginCommand:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
    DtoLoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
        userId:
          type: string
        userName:
          type: string
    DtoReadAllResponse:
      type: object
      properties:
        updated:
          type: integer
          format: int64
    DtoRefreshCommand:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    DtoTransferListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DtoTransferResponse'
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
    DtoTransferResponse:
      type: object
      properties:
        destinationAmount:
          type: string
        destinationCurrency:
          type: string
        destinationFxRate:
          type: string
        failureReason:
          type: string
        feeAmount:
          type: string
        feeCurrency:
          type: string
        fromAccountRef:
          type: string
        message:
          type: string
        sourceAmount:
          type: string
        sourceCurrency:
          type: string
        sourceFxRate:
          type: string
        status:
          type: string
        toAccountName:
          type: string
        toAccountRef:
          type: string
        transactionAmount:
          type: string
        transactionCurrency:
          type: string
        transferId:
          type: string
    DtoUnreadCountResponse:
      type: object
      properties:
        count:
          type: integer
          format: int64
    HandlersErrorResponse:
      type: object
      properties:
        error:
          type: string
        requestId:
          type: string
        traceId:
          type: string